requirements.txt
matplotlib==3.8.3
pandas==2.2.1
Pillow==10.1.0
Pillow==10.2.0
plotly==5.20.0
Requests==2.31.0
streamlit==1.28.2
